requirements.txt
absl-py==0.6.1
astor==0.7.1
backcall==0.1.0
bleach==1.5.0
cycler==0.10.0
decorator==4.3.0
gast==0.2.0
grpcio==1.16.1
h5py==2.8.0
html5lib==0.9999999
hypertools==0.5.1
ipykernel==5.1.0
ipython==7.1.1
ipython-genutils==0.2.0
jedi==0.13.1
jupyter-client==5.2.3
jupyter-core==4.4.0
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
Markdown==3.0.1
matplotlib==3.0.2
mne==0.16.2
numpy==1.15.4
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.7
protobuf==3.6.1
ptyprocess==0.6.0
Pygments==2.2.0
pyparsing==2.3.0
python-dateutil==2.7.5
pyzmq==17.1.2
scikit-learn==0.20.0
scipy==1.1.0
six==1.11.0
sklearn==0.0
tensorboard==1.8.0
tensorflow-gpu==1.8.0
termcolor==1.1.0
tornado==5.1.1
tqdm==4.28.1
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.14.1
